[stable34] fix(files_versions): guard null path in event listeners

+++ apps/files_versions/lib/Listener/FileEventsListener.php
@@ -207,6 +207,9 @@ class FileEventsListener implements IEventListener {
 		}
 
 		$path = $this->getPathForNode($node);
+		if ($path === null) {
+			return;
+		}
 		$result = Storage::store($path);
 
 		// Store the result of the version creation so it can be used in post_write_hook.
@@ -311,6 +314,9 @@ class FileEventsListener implements IEventListener {
 		$node = $this->versionsDeleted[$path];
 		$relativePath = $this->getPathForNode($node);
 		unset($this->versionsDeleted[$path]);
+		if ($relativePath === null) {
+			return;
+		}
 		Storage::delete($relativePath);
 		// If no new version was stored in the FS, no new version should be added in the DB.
 		// So we simply update the associated version.
@@ -324,6 +330,9 @@ class FileEventsListener implements IEventListener {
 	 */
 	public function pre_remove_hook(Node $node): void {
 		$path = $this->getPathForNode($node);
+		if ($path === null) {
+			return;
+		}
 		Storage::markDeletedFile($path);
 		$this->versionsDeleted[$node->getPath()] = $node;
 	}
@@ -344,6 +353,9 @@ class FileEventsListener implements IEventListener {
 
 		$oldPath = $this->getPathForNode($source);
 		$newPath = $this->getPathForNode($target);
+		if ($oldPath === null || $newPath === null) {
+			return;
+		}
 		Storage::renameOrCopy($oldPath, $newPath, 'rename');
 	}
 
@@ -363,6 +375,9 @@ class FileEventsListener implements IEventListener {
 
 		$oldPath = $this->getPathForNode($source);
 		$newPath = $this->getPathForNode($target);
+		if ($oldPath === null || $newPath === null) {
+			return;
+		}
 		Storage::renameOrCopy($oldPath, $newPath, 'copy');
 	}
